Move NotificationListenerService and change message from AlertLevel to String.
