+++ app/src/main/java/org/aurorawatchdevs/aurorawatch/Notification.java
@@ -14,12 +14,12 @@ import org.aurorawatchdevs.aurorawatch.activity.MainActivity;
  * Created by jamesb on 09/06/2015.
  */
 public class Notification {
-    public static void ShowNotification(AlertLevel alertLevel, Context context) {
+    public static void ShowNotification(String alertLevel, Context context) {
         NotificationCompat.Builder mBuilder =
                 new NotificationCompat.Builder(context)
                         .setSmallIcon(R.drawable.aw)
                         .setContentTitle(context.getString(R.string.notification_header))
-                        .setContentText(alertLevel.name().toUpperCase() + " " + context.getString(R.string.notification_body));
+                        .setContentText(alertLevel.toUpperCase() + " " + context.getString(R.string.notification_body));
         // Creates an explicit intent for an Activity in your app
         Intent resultIntent = new Intent(context, MainActivity.class);
